Add horizontal scaling using cluster API
