+++ package.json
@@ -18,7 +18,7 @@
     "start:dev:concurrent": "npx concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run watch-node\"",
     "start:dev": "nodemon --watch \"src/**\" --ext \"ts,json\" --ignore \"src/**/*.test.ts\" --exec \"ts-node src/server.ts\"",
     "start:prod": "npm run build-ts && npm run serve",
-    "start:multi": ""
+    "start:multi": "nodemon --watch \"src/**\" --ext \"ts,json\" --ignore \"src/**/*.test.ts\" --exec \"ts-node src/cluster.ts\""
   },
   "repository": {
     "type": "git",
